+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/ModeCarousel.tsx
@@ -1,70 +1,11 @@
-import { Button } from "@superset/ui/button";
-import { type MotionValue, motion, useMotionValue, useTransform } from "framer-motion";
-import { GitBranch, LayoutList } from "lucide-react";
-import {
-	type ReactNode,
-	useCallback,
-	useEffect,
-	useRef,
-	useState,
-} from "react";
-
-export type SidebarMode = "tabs" | "diff";
-
-const modeIcons: Record<SidebarMode, typeof LayoutList> = {
-	tabs: LayoutList,
-	diff: GitBranch,
-};
-
-const modeLabels: Record<SidebarMode, string> = {
-	tabs: "Tabs",
-	diff: "Diffs",
-};
-
-interface AnimatedBackgroundProps {
-	progress: MotionValue<number>;
-	modeCount: number;
-}
-
-function AnimatedBackground({ progress, modeCount }: AnimatedBackgroundProps) {
-	// Calculate the width of each button (36px = h-9 w-9) + gap (4px = gap-1)
-	const buttonWidth = 36;
-	const gap = 4;
-	const totalButtonWidth = buttonWidth + gap;
-
-	// Transform progress (0-1) to translateX position
-	// For 2 modes: 0 -> 0px, 1 -> 40px (buttonWidth + gap)
-	const translateX = useTransform(
-		progress,
-		[0, modeCount - 1],
-		[0, (modeCount - 1) * totalButtonWidth]
-	);
-
-	return (
-		<motion.div
-			className="absolute h-9 rounded-lg bg-neutral-800/60"
-			style={{
-				width: buttonWidth,
-				x: translateX,
-			}}
-			initial={false}
-			transition={{
-				type: "spring",
-				stiffness: 300,
-				damping: 30,
-			}}
-		/>
-	);
-}
-
-interface ModeCarouselProps {
-	modes: SidebarMode[];
-	currentMode: SidebarMode;
-	onModeSelect: (mode: SidebarMode) => void;
-	children: (mode: SidebarMode, isActive: boolean) => ReactNode;
-	onScrollProgress: (progress: MotionValue<number>) => void;
-	isDragging?: boolean;
-}
+import { useCallback, useRef, useState } from "react";
+import { ModeContent } from "./components/ModeContent";
+import { ModeHeader } from "./components/ModeHeader";
+import { ModeNavigation } from "./components/ModeNavigation";
+import { useModeDetection } from "./hooks/useModeDetection";
+import { useScrollProgress } from "./hooks/useScrollProgress";
+import { useScrollSnap } from "./hooks/useScrollSnap";
+import type { ModeCarouselProps } from "./types";
 
 export function ModeCarousel({
 	modes,
@@ -75,11 +16,7 @@ export function ModeCarousel({
 	isDragging = false,
 }: ModeCarouselProps) {
 	const isInitialMount = useRef(true);
-
 	const currentIndex = modes.findIndex((m) => m === currentMode);
-	const initialProgress = currentIndex >= 0 ? currentIndex : 0;
-	const modeProgress = useMotionValue(initialProgress);
-
 	const [scrollContainer, setScrollContainer] = useState<HTMLDivElement | null>(
 		null,
 	);
@@ -91,127 +28,36 @@ export function ModeCarousel({
 		}
 	}, []);
 
-	// Track scroll position and update motion value
-	useEffect(() => {
-		if (!scrollContainer) return;
+	// Track scroll progress
+	const modeProgress = useScrollProgress({
+		scrollContainer,
+		currentIndex,
+		onScrollProgress,
+	});
 
-		let rafId: number | undefined;
+	// Handle scroll snapping
+	useScrollSnap({
+		scrollContainer,
+		currentIndex,
+		isInitialMount,
+	});
 
-		const updateProgress = () => {
-			const scrollLeft = scrollContainer.scrollLeft;
-			const containerWidth = scrollContainer.offsetWidth;
-			const progress = scrollLeft / containerWidth;
-			modeProgress.set(progress);
-		};
-
-		const handleScroll = () => {
-			// Use requestAnimationFrame for smooth updates
-			if (rafId !== undefined) {
-				cancelAnimationFrame(rafId);
-			}
-			rafId = requestAnimationFrame(updateProgress);
-		};
-
-		scrollContainer.addEventListener("scroll", handleScroll, { passive: true });
-
-		// Initial value
-		updateProgress();
-
-		return () => {
-			scrollContainer.removeEventListener("scroll", handleScroll);
-			if (rafId !== undefined) {
-				cancelAnimationFrame(rafId);
-			}
-		};
-	}, [scrollContainer, modeProgress]);
-
-	// Expose scroll progress to parent
-	useEffect(() => {
-		onScrollProgress(modeProgress);
-	}, [onScrollProgress, modeProgress]);
-
-	// Scroll to current mode when it changes externally
-	useEffect(() => {
-		if (!scrollContainer || currentIndex < 0) return;
-
-		const targetScrollX = currentIndex * scrollContainer.offsetWidth;
-
-		// Only scroll if we're not already at the target position
-		if (Math.abs(scrollContainer.scrollLeft - targetScrollX) > 5) {
-			scrollContainer.scrollTo({
-				left: targetScrollX,
-				behavior: isInitialMount.current ? "auto" : "smooth",
-			});
-		}
-
-		// Mark that initial mount is complete
-		isInitialMount.current = false;
-	}, [currentIndex, scrollContainer]);
-
-	// Detect when user finishes scrolling and update current mode
-	useEffect(() => {
-		if (!scrollContainer || isDragging) return;
-
-		let scrollEndTimer: NodeJS.Timeout | undefined;
-
-		const handleScroll = () => {
-			// Clear existing timeout
-			if (scrollEndTimer) {
-				clearTimeout(scrollEndTimer);
-			}
-
-			// Wait for scroll to settle before updating mode (reduces jitter)
-			scrollEndTimer = setTimeout(() => {
-				const finalScrollLeft = scrollContainer.scrollLeft;
-				const finalContainerWidth = scrollContainer.offsetWidth;
-
-				// Calculate which mode we're closest to and snap to it
-				const finalIndex = Math.round(finalScrollLeft / finalContainerWidth);
-
-				if (
-					finalIndex >= 0 &&
-					finalIndex < modes.length &&
-					modes[finalIndex]
-				) {
-					// Snap to the nearest mode
-					const targetScrollX = finalIndex * finalContainerWidth;
-					if (Math.abs(finalScrollLeft - targetScrollX) > 5) {
-						scrollContainer.scrollTo({
-							left: targetScrollX,
-							behavior: "smooth",
-						});
-					}
-
-					// Update mode if it changed
-					if (modes[finalIndex] !== currentMode) {
-						onModeSelect(modes[finalIndex]);
-					}
-				}
-			}, 150);
-		};
-
-		scrollContainer.addEventListener("scroll", handleScroll, { passive: true });
-
-		return () => {
-			scrollContainer.removeEventListener("scroll", handleScroll);
-			if (scrollEndTimer) {
-				clearTimeout(scrollEndTimer);
-			}
-		};
-	}, [modes, currentMode, onModeSelect, scrollContainer, isDragging]);
+	// Detect mode changes from scrolling
+	useModeDetection({
+		scrollContainer,
+		modes,
+		currentMode,
+		onModeSelect,
+		isDragging,
+	});
 
 	// If only one mode or no modes, disable carousel
 	if (modes.length <= 1) {
 		return (
 			<div className="flex flex-col flex-1 h-full">
 				<div className="flex-1 overflow-y-auto">
-					{/* Header showing current mode */}
-					<div className="px-3 py-2">
-						<span className="text-xs font-medium text-neutral-300">{modeLabels[currentMode]}</span>
-					</div>
-					<div className="px-3">
-						{children(currentMode, true)}
-					</div>
+					<ModeHeader mode={currentMode} />
+					<div className="px-3">{children(currentMode, true)}</div>
 				</div>
 			</div>
 		);
@@ -241,52 +87,28 @@ export function ModeCarousel({
 					{modes.map((mode) => (
 						<div
 							key={mode}
-							className="overflow-y-auto"
 							style={{
-								scrollSnapAlign: "start",
-								scrollSnapStop: "always",
 								width: `${100 / modes.length}%`,
 							}}
 						>
-							{/* Header showing current mode */}
-							<div className="px-3 py-2">
-								<span className="text-xs font-medium text-neutral-300">{modeLabels[mode]}</span>
-							</div>
-							<div className="px-3">
+							<ModeContent
+								mode={mode}
+								isActive={mode === currentMode}
+							>
 								{children(mode, mode === currentMode)}
-							</div>
+							</ModeContent>
 						</div>
 					))}
 				</div>
 			</div>
 
-			{/* Bottom navigation bar - Arc browser style */}
-			<div className="flex items-center justify-center gap-1 px-2 py-2 border-t border-neutral-800/50 bg-neutral-900/50 backdrop-blur-sm">
-				<div className="relative flex items-center gap-1">
-					{/* Animated background indicator */}
-					<AnimatedBackground progress={modeProgress} modeCount={modes.length} />
-
-					{modes.map((mode) => {
-						const Icon = modeIcons[mode];
-						const isActive = mode === currentMode;
-
-						return (
-							<Button
-								key={mode}
-								variant="ghost"
-								size="sm"
-								onClick={() => onModeSelect(mode)}
-								className={`relative z-10 h-9 w-9 rounded-lg transition-colors duration-200 ${isActive
-									? "text-neutral-100"
-									: "text-neutral-400 hover:text-neutral-300"
-									}`}
-							>
-								<Icon className="w-4 h-4" />
-							</Button>
-						);
-					})}
-				</div>
-			</div>
+			{/* Bottom navigation bar */}
+			<ModeNavigation
+				modes={modes}
+				currentMode={currentMode}
+				onModeSelect={onModeSelect}
+				scrollProgress={modeProgress}
+			/>
 		</div>
 	);
 }
