+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/components/AnimatedBackground/index.ts
@@ -1,0 +1,2 @@
+export { AnimatedBackground } from "./AnimatedBackground";
+
