+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/components/ModeContent/index.ts
@@ -1,0 +1,2 @@
+export { ModeContent } from "./ModeContent";
+
