+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/components/ModeHeader/ModeHeader.tsx
@@ -1,0 +1,17 @@
+import { modeLabels } from "../../constants";
+import type { SidebarMode } from "../../types";
+
+interface ModeHeaderProps {
+	mode: SidebarMode;
+}
+
+export function ModeHeader({ mode }: ModeHeaderProps) {
+	return (
+		<div className="px-3 py-2">
+			<span className="text-xs font-medium text-neutral-300">
+				{modeLabels[mode]}
+			</span>
+		</div>
+	);
+}
+
