+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/components/ModeHeader/index.ts
@@ -1,0 +1,2 @@
+export { ModeHeader } from "./ModeHeader";
+
