+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/components/ModeNavigation/index.ts
@@ -1,0 +1,2 @@
+export { ModeNavigation } from "./ModeNavigation";
+
