+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/constants.ts
@@ -1,0 +1,13 @@
+import { GitBranch, LayoutList } from "lucide-react";
+import type { SidebarMode } from "./types";
+
+export const modeIcons: Record<SidebarMode, typeof LayoutList> = {
+	tabs: LayoutList,
+	diff: GitBranch,
+};
+
+export const modeLabels: Record<SidebarMode, string> = {
+	tabs: "Tabs",
+	diff: "Diffs",
+};
+
