+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/index.ts
@@ -1,2 +1,2 @@
-export { ModeCarousel, type SidebarMode } from "./ModeCarousel";
-
+export { ModeCarousel } from "./ModeCarousel";
+export type { SidebarMode, ModeCarouselProps } from "./types";
