+++ apps/desktop/src/renderer/screens/main/components/Sidebar/components/ModeCarousel/types.ts
@@ -1,0 +1,13 @@
+import type { MotionValue, ReactNode } from "react";
+
+export type SidebarMode = "tabs" | "diff";
+
+export interface ModeCarouselProps {
+	modes: SidebarMode[];
+	currentMode: SidebarMode;
+	onModeSelect: (mode: SidebarMode) => void;
+	children: (mode: SidebarMode, isActive: boolean) => ReactNode;
+	onScrollProgress: (progress: MotionValue<number>) => void;
+	isDragging?: boolean;
+}
+
